feat: Add Discover RPC to FilesService
This commit adds a new RPC called Discover to the FilesService. It also includes changes to the files.proto and Files.fs files. The files.proto file now includes a new message called FilesServiceDiscoverRequest and a corresponding response message called FilesServiceDiscoverResponse. The Files.fs file has been updated with the necessary code for handling the new request and response messages.